requirements.txt
PyQt6
pymongo
validators
python-dotenv
Flask